Feature/stuck fixup (#104) (#105)
* wbc-4g-usb: fix deinit hook to prevent boot stuck

* .service: add startup timeout

+++ modules/wbc-4g-usb.sh
@@ -5,7 +5,7 @@ hook_module_add() {
 }
 
 hook_module_del() {
-	[[ -z "$NO_RESTART_SERVICE" ]] && {
+	if [[ -z "$NO_RESTART_SERVICE" ]]; then
 		systemctl stop wb-gsm || true
-	}
+	fi;
 }
